backend/requirements.txt
# FastAPI 核心框架和ASGI服务器
# FastAPI core framework and ASGI server
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# LangChain 生态系统 - LLM应用开发框架
# LangChain ecosystem - LLM application development framework
langchain==0.3.25
langchain-anthropic==0.3.4
langgraph==0.4.6
langchain-community==0.3.24
langchain-core==0.3.25
langsmith==0.2.30

# HTTP 客户端和异步编程
# HTTP client and async programming
httpx==0.28.1
aiofiles==24.1.0

# 定时任务调度器
# Task scheduler
apscheduler==3.11.0

# Azure 云服务 SDK
# Azure cloud services SDK
azure-search-documents==11.6.0
azure-storage-blob==12.23.1

# 数据库相关
# Database related
psycopg[binary]==3.2.3
sqlalchemy==2.0.36
alembic==1.14.0

# OAuth 和认证
# OAuth and authentication
authlib==1.3.2
itsdangerous==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
PyJWT==2.10.1
email-validator==2.2.0

# 数据验证和序列化
# Data validation and serialization
pydantic==2.10.3
pydantic-settings==2.7.0

# 环境变量管理
# Environment variable management
python-dotenv==1.0.1

# 图表和数据处理
# Charts and data processing
chart-studio==1.1.0
pandas==2.2.3
numpy==2.2.1
scikit-learn==1.6.0

# 文档和PDF处理
# Document and PDF processing
pypdf==5.2.0
python-docx==1.1.2

# Redis 缓存 (可选)
# Redis cache (optional)
redis==5.2.1

# 监控和日志
# Monitoring and logging
sentry-sdk[fastapi]==2.21.0

# 开发工具
# Development tools
pytest==8.3.4
pytest-asyncio==0.25.0
black==24.10.0
isort==5.13.2
flake8==7.1.1 